Lambda/deployment/requirements.txt
onnxruntime
numpy
Pillow